requirements.txt
absl-py==2.3.1
altair==5.5.0
anyio==4.9.0
apturl==0.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcrypt==3.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==6.1.0
Caffe==0.1.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.0.3
cmdstanpy==1.2.4
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.3.2
convertdate==2.4.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
Cython==0.29.21
dbus-python==1.2.18
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
ephem==4.2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fasteners==0.14.1
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
fqdn==1.5.1
fsspec==2025.7.0
future==0.18.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
hijri-converter==2.3.2.post1
hijridate==2.3.0
holidays==0.25
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.3
imbalanced-learn==0.13.0
imblearn==0.0
importlib-metadata==4.6.4
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.10.0
keyring==23.5.0
kiwisolver==1.4.8
korean-lunar-calendar==0.3.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
LunarCalendar==0.0.9
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
namex==0.1.0
narwhals==1.48.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
notebook==7.4.4
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
overrides==7.7.0
packaging==25.0
pandas==2.2.2
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
patsy==1.0.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.8
plotly==6.2.0
pmdarima==2.0.4
prometheus_client==0.22.1
prompt_toolkit==3.0.51
prophet==1.1.5
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydeck==0.9.1
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMeeus==0.5.12
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pystan==2.19.1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-json-logger==3.3.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==27.0.0
referencing==0.36.2
reportlab==3.6.8
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.26.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
ssh-import-id==5.11
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.5
streamlit==1.48.1
sympy==1.14.0
systemd-python==234
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
terminado==0.18.1
Theano==1.0.5
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.31.2
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
xdg==5
xgboost==3.0.2
xkit==0.0.0
zipp==1.0.0
